requirements.txt
playwright==1.40.0
playwright-stealth==1.0.6
pandas==2.0.0
openpyxl==3.1.2
python-dotenv==1.0.0
rich==13.7.0
typer==0.9.0
